requirements.txt
streamlit
firebase-admin
google-generativeai
streamlit-folium
pandas
pillow
